+++ src/main/java/com/example/PrescriptionCRUD/services/PatientService.java
@@ -5,7 +5,10 @@ import com.example.PrescriptionCRUD.repositories.PatientRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -18,13 +21,43 @@ public class PatientService {
     }
 
     public void addPatient(Patient patient){
-        patientRepository.save(patient);
+        Optional<Patient> patientOptional = patientRepository.findByPesel(patient.getPesel());
+        if (patientOptional.isPresent()){
+            patientRepository.save(patient);
+        } else {
+            throw new IllegalStateException("Patient already exists in database.");
+        }
     }
 
     public void deletePatient(Long patientId){
-        patientRepository.deleteById(patientId);
+        Optional<Patient> patientOptional = patientRepository.findById(patientId);
+        if (patientOptional.isPresent()){
+            patientRepository.deleteById(patientId);
+        } else {
+            throw new IllegalStateException("This patient is not present in database.");
+        }
     }
 
+    @Transactional
+    public void updatePatient(Long patientId, String name, String lastName){
+        Optional<Patient> patientOptional = patientRepository.findById(patientId);
+        if (patientOptional.isPresent()){
+            Patient patient = patientOptional.get();
+            if (name != null && !name.isEmpty() && !Objects.equals(patient.getName(), name)){
+                patient.setName(name);
+            } else {
+                throw new IllegalStateException("Invalid name provided!");
+            }
+
+            if (lastName != null && !lastName.isEmpty() && !Objects.equals(patient.getLastName(), lastName)) {
+                patient.setLastName(lastName);
+            } else {
+                throw new IllegalStateException("Invalid last name provided!");
+            }
+        } else {
+            throw new IllegalStateException("This patient does not exist in database! ");
+        }
+        }
 
 }
